backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic-settings>=2.6,<3.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
alembic>=1.14,<2.0
pytest>=8.3,<9.0
pandas>=2.2,<3.0
numpy>=2.1,<3.0
scikit-learn>=1.5,<2.0
joblib>=1.4,<2.0
